+++ src/sharc-container.js
@@ -26,7 +26,7 @@
  * container.load();
  * ```
  *
- * @version 0.7.8
+ * @version 0.7.9
  */
 
 'use strict';